# fix: core icon performance and a11y