--- a/docs/core/cwc-icon.md
+++ b/docs/core/cwc-icon.md
@@ -20,7 +20,6 @@ ClarityIcons.addIcons(userIcon);
 |----------|-----------|----------|--------------|--------------------------------------------------|
 | `dir`    | `dir`     | `string` | **required** | Takes a directional value (up\|down\|left\|right) that rotates the icon 90° with the<br />top of the icon pointing in the specified direction. |
 | `flip`   | `flip`    | `string` | **required** | Takes an orientation value (horizontal\|vertical) that reverses the orientation of the<br />icon vertically or horizontally. |
-| `id`     | `id`      | `string` | **required** | If present, determines the id of the icon. Uses a generated unique id otherwise. |
 | `shape`  | `shape`   | `string` |              | Changes the svg glyph displayed in the icon component. Defaults to the 'unknown' icon if<br />the specified icon cannot be found in the icon registry. |
 | `size`   | `size`    | `string` |              | Apply numerical width-height or a t-shirt-sized CSS classname |
 | `title`  | `title`   | `string` | **required** | If present, customizes the aria-label for the icon for accessibility. |