## PR Checklist

Please check if your PR fulfills the following requirements:

* [x] Tests for the changes have been added (for bug fixes / features)
* [ ] Docs have been added / updated (for bug fixes / features)
* [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

* [x] Bugfix
* [ ] Feature
* [ ] Code style update (formatting, local variables)
* [x] Refactoring (no functional changes, no api changes)
* [ ] Build related changes
* [ ] CI related changes
* [ ] Documentation content changes
* [ ] clarity.design website / infrastructure changes
* [ ] Other... Please describe:

## What is the new behavior?
### 1. Improved icon demo performance
  - Renders demo page (416 icons) around ~630-750ms compared to 1.1s - 1.2s on docs icon demo page built in prod mode on 2018 macbook pro.
  - Remove use of getters and replace with form value changes observable. This prevents recalculation for each icon during each change detection cycle.
  - Load icons immediately and create index object inline to prevent multiple renders/change detection cycles when angular component is created.
  - Use CSS to show and hide icons to prevent large amount of components being created/destroyed when filtering/searching
  - Remove table layout to reduce number of DOM nodes improving render times

### 2. Improved icon element render performance
  - Single cwc-icon uses ~0.64ms of JS execution time vs current clr-icon at ~1.5ms (2018 macbook pro)
  - remove unnecessary id reflection, ids for a11y labeling are internal and unique so syncing was not needed for a11y
  - only write to svg and create SR label if title is set
  - default svg template to have `role="img"` instead of writing on creation
  - remove unnecessary `<div>` wrapper to reduce DOM elements

### 3. Fix a11y behavior with screen readers.
  - Icons default decorative images unless otherwise set with `title`. When set with title a SR only message will be rendered with title value. Icon should be described by the context of usage such as text within a button or a `aria-label` on a button. Only if icon is represented as stand alone content is the `title` needed. https://clarity.design/icons/accessibility 

## Does this PR introduce a breaking change?

* [ ] Yes
* [x] No

<!-- If this PR contains a breaking change, please describe the impact and migration path for existing applications below. -->

## Other information